wip; cp;
all tests pass except the inline adapter 2 sandbox test

the problem is that each sandbox is calling `oasis.connect` on its own
oasis instance, *but* the internals of oasis rely on the global `oasis`
instance, which will turn out to be the most recently created one.

grep this commit for TODOs.